refactor(bridge): extract SourceMessageHandler
